+++ test/DateTimesTest.js
@@ -3,7 +3,7 @@
 "use strict";
 var chai = require("chai");
 var DateTimes = require("../DateTimes");
-var DateConstants = DateTimes.DateConstants, Dates = DateTimes.Dates, Timestamps = DateTimes.Timestamps, DotNetJsonDate = DateTimes.DotNetJsonDate;
+var DateConstants = DateTimes.DateConstants, Dates = DateTimes.Dates, DateUtil = DateTimes.DateUtil, Timestamps = DateTimes.Timestamps, DotNetJsonDate = DateTimes.DotNetJsonDate;
 var asr = chai.assert;
 suite("DateTimes", function DateTimesTest() {
     function testDotNetJson(asr) {
@@ -92,5 +92,21 @@ suite("DateTimes", function DateTimesTest() {
         asr.equal("01/27/2000 02:14 a.m.", res1);
         var res1 = Timestamps.toDisplayDateTime(new Date(2000, 0, 27, 23, 8).getTime(), false);
         asr.equal("01/27/2000 11:08 p.m.", res1);
+    });
+    test("Timestamp.toDisplayTime", function toDisplayTimeTest() {
+        var res1 = Timestamps.toDisplayTime(new Date(2000, 0, 27, 0, 0).getTime());
+        asr.equal("12:00 a.m.", res1);
+        var res1 = Timestamps.toDisplayTime(new Date(2000, 0, 27, 2, 14).getTime());
+        asr.equal("02:14 a.m.", res1);
+        var res1 = Timestamps.toDisplayTime(new Date(2000, 0, 27, 23, 8).getTime());
+        asr.equal("11:08 p.m.", res1);
+    });
+    test("Timestamp.isSameDate", function isSameDateTest() {
+        var res1 = DateUtil.isSameDate(new Date(2000, 0, 27), new Date(2000, 0, 28));
+        asr.equal(false, res1);
+        var res1 = DateUtil.isSameDate(new Date(2000, 0, 27, 2, 14), new Date(2000, 0, 27, 23));
+        asr.equal(true, res1);
+        var res1 = DateUtil.isSameDate(new Date(2000, 0, 27, 23, 8), new Date(2000, 0, 27));
+        asr.equal(true, res1);
     });
 });